User commit

+++ src/main/java/com/example/demo/controller/ValidationController.java
@@ -21,7 +21,7 @@ public ValidationEntity dopost( @Valid @RequestBody ValidationEntity stuu){
 
 }
 @GetMapping("/getidd{id}")
-public ValidationEntity getdataid(@PathVariable int id){
+public ValidationEntity getdataid(@PathVariable Long id){
     return ser.findDataa(id);
 }
