+++ src/main/java/com/example/demo/service/ValidationService.java
@@ -4,7 +4,7 @@ import com.example.demo.entity.ValidationEntity;
 public interface ValidationService{
     ValidationEntity postdata(ValidationEntity stuu);
    // List<ValidationEntity>getAllData();
-    ValidationEntity findDataa(int id);
+    ValidationEntity findDataa(Long id);
     
 }
